example/requirements/base.txt
django==4.2
skip-django-germanium @ git+https://github.com/skip-pay/germanium@tda/chore/django_bump
coveralls
six==1.10.0
flake8
-e ../
